requirements.txt
autogen-agentchat
autogen-core
autogen-ext[openai]
autogen-ext[docker]
docker
python-dotenv
tiktoken
streamlit
asyncio